backend/requirements.txt
ultralytics
Pillow
numpy